backend/data_processing/requirements.txt
fastapi==0.121.3
uvicorn[standard]==0.38.0
pydantic==2.12.4
SQLAlchemy==2.0.44
asyncpg==0.30.0
httpx==0.28.1
numpy<2.0
pillow
opencv-python-headless
scikit-image